pyproject.toml
[project]
name = "financial-data-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.6.0",
    "fastapi>=0.115.12",
    "hf-xet>=1.0.0",
    "nltk>=3.9.1",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "torch>=2.6.0",
    "transformers>=4.51.0",
    "uvicorn>=0.34.0",
]
